♻️  soundMeter 코드 수정
