+++ client/src/components/room/in-room-user-box.tsx
@@ -7,7 +7,7 @@ import {
 } from 'react-icons/fi';
 
 import { getUserInfo } from '@api/index';
-import SoundMeter from '@src/utils/test';
+import SoundMeter from '@src/utils/voice';
 import { Link } from 'react-router-dom';
 import { InRoomUserBoxStyle, InRoomUserMicDiv, UserBox } from './style';
 
@@ -84,7 +84,7 @@ export const InRoomUserBox = React.forwardRef<HTMLVideoElement, IParticipant>(
       if (!props.stream || !props.isMicOn) return;
       const soundMeter = new SoundMeter(audioCtxRef.current);
       let meterRefresh: any = null;
-      soundMeter.connectToSource(props.stream, (e: any) => {
+      soundMeter.connectToSource(props.stream, () => {
         meterRefresh = setInterval(() => {
           const num = Number(soundMeter.instant.toFixed(2));
           if (num > 0.02 && myRef.current) {
@@ -98,6 +98,7 @@ export const InRoomUserBox = React.forwardRef<HTMLVideoElement, IParticipant>(
       // eslint-disable-next-line consistent-return
       return () => {
         clearInterval(meterRefresh);
+        soundMeter.stop();
       };
     }, [props.isMicOn]);
